feat: use unix timestamp in generateMessage

+++ structs/index.go
@@ -3,5 +3,5 @@ package structs
 // QueueMessage defines the message to be sent to server
 type QueueMessage struct {
 	Title     string
-	Timestamp int32
+	Timestamp int64
 }
